+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Master
+
+- Use `$stderr` instead of `STDERR` to allow redirection.
+ 
 ## 1.4.1 - 5 October, 2012
 
 - [#334][] Extend `:tmux` notifier with use of `tmux display-message` and options to configure them. ([@matthijsgroen][])